Timeline-Trendings : Optimize Timeline's functions and adding Trendings page
