+++ client/src/timeline/component/subscribe/subscriptionTimelineGrid.js
@@ -2,73 +2,13 @@
 
 import React, { useEffect, useState } from "react";
 import axios from "axios";
+import {GetTimeElapsed, TimeOfVideo} from "../../functions/VideoTiming";
 
 import CheckSession from "../../../session"
 //const { isLoggedIn, userId } = CheckSession();
 
 var userId = 1;
 
-// Fonction pour calculer le temps écoulé depuis la date d'upload
-function getTimeElapsed(uploadDateTime) {
-  const uploadDate = new Date(uploadDateTime);
-  const currentDate = new Date();
-
-  const elapsedMilliseconds = currentDate - uploadDate;
-  const elapsedSeconds = Math.floor(elapsedMilliseconds / 1000);
-  const elapsedMinutes = Math.floor(elapsedSeconds / 60);
-  const elapsedHours = Math.floor(elapsedMinutes / 60);
-  const elapsedDays = Math.floor(elapsedHours / 24);
-
-  if (elapsedDays > 0) {
-    return `${elapsedDays} days`;
-  } else if (elapsedHours > 0) {
-    return `${elapsedHours} hours`;
-  } else if (elapsedMinutes > 0) {
-    return `${elapsedMinutes} minutes`;
-  } else {
-    return `${elapsedSeconds} seconds`;
-  }
-}
-
-function timeOfVideo(totalSeconds) {
-  var hours;
-  var minutes;
-  var seconds;
-  hours = Math.floor(totalSeconds / 3600);
-  totalSeconds %= 3600;
-  minutes = Math.floor(totalSeconds / 60);
-  seconds = totalSeconds % 60;
-  var result = "";
-  var resultSeconds = "";
-  var resultMinutes = "";
-  var resultHours = "";
-
-  // Seconds
-  if (seconds < 10) {
-    resultSeconds = `0${seconds}`;
-  } else {
-    resultSeconds = `${seconds}`;
-  }
-  // Minutes
-  if (minutes < 10) {
-    resultMinutes = `0${minutes}:`;
-  } else {
-    resultMinutes = `${minutes}:`;
-  }
-  // Hours
-  if (hours > 0) {
-    if (hours < 10) {
-      resultHours = `0${hours}:`;
-    } else {
-      resultHours = `${hours}:`;
-    }
-  } else {
-    resultHours = ``;
-  }
-
-  result = `${resultHours}${resultMinutes}${resultSeconds}`;
-  return result;
-}
 
 export default function GridSubscriptionTimeLine() {
   // Get the informations of the SQL Request by the URL
@@ -107,7 +47,7 @@ export default function GridSubscriptionTimeLine() {
 
   for (var i = 0; i < videosInfos.length; i++) {
     var date = videosInfos[i]["upload_date_time"];
-    var videoLenght = timeOfVideo(videosInfos[i]["video_duration"]);
+    var videoLenght = TimeOfVideo(videosInfos[i]["video_duration"]);
     indents.push(
       <div key={i} className="max-w-[25%] h-auto mb-2">
         <a href={`/watch?video_id=${videosInfos[i]["id"]}`}>
@@ -133,7 +73,7 @@ export default function GridSubscriptionTimeLine() {
               </h4>
               <h4 className="text-gray text-[90%]">
                 {videosInfos[i]["number_view"]} views -{" "}
-                {getTimeElapsed(videosInfos[i]["upload_date_time"])} ago
+                {GetTimeElapsed(videosInfos[i]["upload_date_time"])} ago
               </h4>
             </div>
           </div>
