+++ client/src/timeline/component/timelineRightSide.js
@@ -3,70 +3,7 @@
 import React, { useEffect, useState } from "react";
 import axios from 'axios';
 import {SetScores} from "../functions/AdvancedTimelineCalculator.js";
-
-// Function to calculate time elapsed since upload date
-function getTimeElapsed(uploadDateTime) {
-  const uploadDate = new Date(uploadDateTime);
-  const currentDate = new Date();
-
-  const elapsedMilliseconds = currentDate - uploadDate;
-  const elapsedSeconds = Math.floor(elapsedMilliseconds / 1000);
-  const elapsedMinutes = Math.floor(elapsedSeconds / 60);
-  const elapsedHours = Math.floor(elapsedMinutes / 60);
-  const elapsedDays = Math.floor(elapsedHours / 24);
-
-  if (elapsedDays > 0) {
-    return `${elapsedDays} days`;
-  } else if (elapsedHours > 0) {
-    return `${elapsedHours} hours`;
-  } else if (elapsedMinutes > 0) {
-    return `${elapsedMinutes} minutes`;
-  } else {
-    return `${elapsedSeconds} seconds`;
-  }
-}
-
-
-// Function that translates time in seconds into HH:MM:SS format
-function timeOfVideo(totalSeconds) {
-  var hours;
-  var minutes;
-  var seconds;
-  hours = Math.floor(totalSeconds / 3600);
-  totalSeconds %= 3600;
-  minutes = Math.floor(totalSeconds / 60);
-  seconds = totalSeconds % 60;
-  var result = "";
-  var resultSeconds = "";
-  var resultMinutes = "";
-  var resultHours = "";
-
-  // Seconds
-  if (seconds < 10) {
-    resultSeconds =  `0${seconds}`
-  } else {
-    resultSeconds =  `${seconds}`
-  }
-  // Minutes
-  if (minutes < 10) {
-    resultMinutes =  `0${minutes}:`
-  } else {
-    resultMinutes =  `${minutes}:`
-  }
-  // Hours
-  if (hours > 0) {
-    if (hours < 10) {
-      resultHours =  `0${hours}:`
-    } else {
-      resultHours =  `${hours}:`
-    }
-  } else {
-    resultHours =  ``
-  }
-  
-  result = `${resultHours}${resultMinutes}${resultSeconds}`
-  return result;
-}
+import {GetTimeElapsed, TimeOfVideo} from "../functions/VideoTiming.js";
 
 export default function TimelineRightSide() {
   var [videosInfos, setVideosInfos] = useState([]);
@@ -88,7 +25,7 @@ export default function TimelineRightSide() {
   var indents = [];
   for (var i = 0; i < videosInfos.length; i++) {
     var date = videosInfos[i]["upload_date_time"];
-    var videoLenght = timeOfVideo(videosInfos[i]["video_duration"])
+    var videoLenght = TimeOfVideo(videosInfos[i]["video_duration"])
 
     indents.push(
       <div key={i} className="h-auto mb-2 ">
@@ -113,7 +50,7 @@ export default function TimelineRightSide() {
                 {videosInfos[i]["pseudo"]}
               </h4>
               <h4 className="text-gray text-[90%]">
-                {videosInfos[i]["number_view"]} views - {getTimeElapsed(videosInfos[i]["upload_date_time"])} ago
+                {videosInfos[i]["number_view"]} views - {GetTimeElapsed(videosInfos[i]["upload_date_time"])} ago
               </h4>
               <p className="font-bold text-purple-700">Score: {videosInfos[i]["score"]}</p>
             </div>
