+++ client/src/timeline/functions/VideoTiming.js
@@ -1,0 +1,65 @@
+import React, { useEffect, useState } from "react";
+import axios from "axios";
+
+// Function to calculate time elapsed since upload date
+export function GetTimeElapsed(uploadDateTime) {
+    const uploadDate = new Date(uploadDateTime);
+    const currentDate = new Date();
+  
+    const elapsedMilliseconds = currentDate - uploadDate;
+    const elapsedSeconds = Math.floor(elapsedMilliseconds / 1000);
+    const elapsedMinutes = Math.floor(elapsedSeconds / 60);
+    const elapsedHours = Math.floor(elapsedMinutes / 60);
+    const elapsedDays = Math.floor(elapsedHours / 24);
+  
+    if (elapsedDays > 0) {
+      return `${elapsedDays} days`;
+    } else if (elapsedHours > 0) {
+      return `${elapsedHours} hours`;
+    } else if (elapsedMinutes > 0) {
+      return `${elapsedMinutes} minutes`;
+    } else {
+      return `${elapsedSeconds} seconds`;
+    }
+  }
+  
+  // Function that translates time in seconds into HH:MM:SS format
+export function TimeOfVideo(totalSeconds) {
+    var hours;
+    var minutes;
+    var seconds;
+    hours = Math.floor(totalSeconds / 3600);
+    totalSeconds %= 3600;
+    minutes = Math.floor(totalSeconds / 60);
+    seconds = totalSeconds % 60;
+    var result = "";
+    var resultSeconds = "";
+    var resultMinutes = "";
+    var resultHours = "";
+  
+    // Seconds
+    if (seconds < 10) {
+      resultSeconds =  `0${seconds}`
+    } else {
+      resultSeconds =  `${seconds}`
+    }
+    // Minutes
+    if (minutes < 10) {
+      resultMinutes =  `0${minutes}:`
+    } else {
+      resultMinutes =  `${minutes}:`
+    }
+    // Hours
+    if (hours > 0) {
+      if (hours < 10) {
+        resultHours =  `0${hours}:`
+      } else {
+        resultHours =  `${hours}:`
+      }
+    } else {
+      resultHours =  ``
+    }
+    
+    result = `${resultHours}${resultMinutes}${resultSeconds}`
+    return result;
+  }
