+++ client/src/timeline/pages/trends.js
@@ -1,20 +1,31 @@
 //  PAGE DISPLAYING THE TRENDINGS TIMELINE
 
+import NonDisplayedBurgerMenu from "../component/nonDisplayedBurgerMenu";
+import TrendingsTimeLine from "../component/TrendingsTimeline";
+import TimelineRightSide from "../component/timelineRightSide";
+import "../styles/Timeline.css";
 import React, { useEffect, useState } from "react";
 
-export default function Trends() {
+export default function Mainpage() {
 
   /* Page title (displayed on tab) */
   useEffect(() => {
-    document.title = "Trends - CodingTube";
+    document.title = "Trendings - CodingTube";
   }, []);
   return (
     <>
+
       {/* TITLES */}
-      <div>
-        <h1 className="text-3xl font-bold underline">
-          Here is the trending page
-        </h1>
+      <div className="ml-24 mt-8">
+        <h1 className="text-3xl font-bold">Top Trendings</h1>
+      </div>
+
+      {/* PAGE CONTENT */}
+      <div className="flex">
+        <NonDisplayedBurgerMenu /> {/* Burger menu unfolded */}
+        <div class="flex ml-[3%] flex-col content-center">
+          <TrendingsTimeLine /> {/* Showing Trendings Timeline */}
+        </div>
       </div>
     </>
   );
