+++ client/src/timeline/styles/Timeline.css
@@ -14,3 +14,9 @@
     height: auto;
     border-radius: 10px;
 }
+
+.thumbnail-trendings-list {
+    width: 300px;
+    height: auto;
+    border-radius: 10px;
+}
